debug bus error
We were receiving a bus error when running stress2.hs. Ultimately,
this occurs because we do not have the GC hooked up yet. We are
getting a heap overflow, telling the RTS to GC, but we aren't.
So the allocation is overflowing into space that is unknown and
when we try to access it, we get the bus error.