fix bus error when expanding nursery
The nurseries were getting screwed up when we expand them causing
bus errors to occur. This change fixes that. We should now be
able to allocate as much as we want and the nursery will grow with
it -- although this is not the desired end behavior (since we need
to impose a limit).