implement hierarchical scheduling primitives
The idea behind hierarchical scheduling is to partition the CPU
such that each thread has a bounded amount of time to run. A thread
is able to give up CPU time in order to fork. Additionally, it can
be given CPU time back when a child thread finishes.

This commit implements tick-based scheduling where each thread can
run for a different amount of ticks, as specified in the TSO structure.
If a thread completes, it is put in the new ThreadBusyWait state. A
thread will remain in this state until killed. In the state, it will
continue to be scheduled for its time slice.

The rts also exposes a set of primitives to add low level CPU management
to Haskell code. An example use of this can be seen in `attacks/common.hs`.
The rts exposes functions for threads to add CPU time (to themselves),
relinquish time, and get their own time.