implement functional hierarchical scheduling
This is a base implementation of hierarchical scheduling.
Hierarchical scheduling maintains the parent/child relationship
of threads within the run queue. Rather than the linked list
run queue of the rts, each thread can also link to children,
a parent, and other threads from the same parent.

Additionally, each H thread is run for a specific number of ticks.
This is determined by the application and can be changed by the
application as well.

It is important to note that since threads are hierarchical, when a
thread is killed, all of its children are killed as well. This is
an inherent property of the hierarchical queue.

There are still some major changes to make to this:
(1) Expose a wait feature that uses ticks (as opposed to IO.threadDelay)
(2) Cross core kills
(3) Implement a sane LIO API
(4) Test/integrate with Frankie
(5) Perf measurements

Things to be weary of/test/look out for:
(1) Synchronization primitives---when threads are transitively killed,
they do not go through the ThreadKilled process. Rather they wipe off
the face of the earth. This could be unsafe if one thread is blocked
on a transitively killed thread as the thread would freeze.