hthread checkpoint
Right now, we can compile with stage-1 and execute some programs that
fork threads. However, it does not work when linked to stage-2 and
used to compile a program. The error is that when releasing a
capability, it thinks there is always work to do and creates
pthreads in an infinite loop. The OS eventually tells the rts
that it can't create any more and the program crashes.