fix iomanager infinite loop bug
The IOManager did not check whether the TSOs it created were the
first ones on a capability. This caused an infinite loop where
the capability would always yield instead of execute code. It was
fixed by adding a check to scheduleWaitThread to see if the TSO
being scheduled is the first on the run queue.

This commit is the first where we can compile and execute all of
the nofib test suite.