add default ticks as +RTS argument
Default ticks are the number of ticks that each TSO will be
scheduled for. This value defaults to zero. This is equivalent
to context switching normally. If set to a value, each thread will
context switch after the set amount of ticks. This value, in the
future, will be able to be specified via a new fork function or
via a setter function on a per thread basis.