fix deadlock when timer stops
The timer is stopped during a GC. If a busy wait occurred when a
GC was starting, the system would freeze because the waiting thread
would not get the signal while it was looping.

This change adds a check to busy wait so that it can tell if the
timer is actually going. If so, the busy wait continues. If not,
the waiting tso is remembered and jumps back to the scheduling
loop to GC. Afterwhich, the remembered TSO is scheduled to busy
wait again.