start timeout implemenation
Hierarchical threads can be configured to timeout. This is
configurable via the command line or the Haskell API. Although this
implementation works, it is dangerous to actually use. There is no
cleanup when killing the threads so it is very easy to deadlock
the system. Additionally, there is no mechanism in place to
transitively kill all children and reclaim ticks.